add a new memory pool : datastack pool
beware that a realloc using that pool can invalidate a previously ds_get-ed
memory area.

this is not a surprise, since it's coherent with data-stacks in general, but
the fact that we use it through the data-stack pool may hide that fact a bit


This pool is intended to be used for temporary buffers in the blob_*
functions.


btw, always consider that when you ds_get sth, if you don't ds_tie it before
any following function call, and that you hope using this area after this
function call, you *are* wrong.