remove old 'trace' function, and use e_debug everywhere.
add noreturn attributes to the e_fatal/e_panic.
make __attribute__(...) expand to nothing instead of redefining each macro on non GNUC compilers.
(approx: check if GNUC is >= 3.x, who still uses gcc 2.95 anyway ?)

Note for self: there is a couple of nice attributes to investigate for our libcommon (like nonnull)
because it's a good way to document the API...