add a nice way to enable debug through the env var IS_DEBUG:
you can have a blank separated list of <fname> or <funcspec> where:

 <fname> is a c module name (with the leading .c).
 <funcspec> is a c module slash ('/') a function name you want to watch.

e.g.:

  IS_DEBUG="blob.c worker.c/worker_read"

will print *any* debug in blob.c and *any* debug in the function
worker_read in worker.c.

the code is pretty naïve, and allocates memory it never free right now.

TODO: add to the spec a level so that we can force a per module/function
      debug level.