mmfile_truncate enhancements.
rework things in the mmfile_truncate function so that it borks the mmfile
in a very few (rare) occasions.

also let the resultant mmfile be a valid one, rather than a dangling.

Use a linux dependent behaviour of munmap to unmap unused parts of the
memory if the file shrinks. guard that with __linux__ ifdefs as I'm not sure
it works on cygwin or other unixes.