remove dealt with and inacurate comments.
indeed the nullified magic is not good, but the dirty bit is not really
better, as it causes exactly the same kind of problems upon opening anyway.

put all that code in #if 0 blocks until we get a better idea: we have to
make a difference between a file that is already mapped somewhere in
write mode and a file that is uncleanly closed.

Signed-off-by: pierre habouzit <pierre.habouzit@intersec.eu>