Have the mmfile embed its own locking functions.
  Use them for write operations like truncate or close.

  Have 3 pointers: lock, unlock, destroy. The mutex is not held in the
structure on purpose, as if we want to use it across multiple processes, we
will put it in the mapped file or in shared memory and use
pthread_*_setpshared functions to share it, it cannot live in the structure.

  The lock is intended to be a rwlock-like lock, and write lock is used in
very rare cases, hence should not generate too much bottlenecks.

NB: the possibility to have the lock in the map has to be evaluated, even if
    it's not useful in our current perspective anyways.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>