Add CYCLIC_MIN/CYCLIC_MAX macros.
  Those are intended to guess the MAX value of integers who wrap around
continuously. It must be used in environment that satisfy this requirement:

  for any value v1 and v2, emitted in that order, and that are still alive
at a date t, then (v2 - v1) < (v1 - v2). Said in natural language, it means
that when it wraps and that it's stored in uint32_t's, when it becomes 0
again, nobody has any references on a value less or equal to (UINT_MAX/2 + 1).

  IOW: it's used to store monotonic clocks, and that you don't need to
remember too far in the past. Note that for a 32bits unsigned integer, if you
tick every 1/1000 seconds, then you can acurately describe an event that
happened 24 days ago. or for a tick every usec, it's 14 hours.

  Example:

  for a uint8_t:
  * CYCLE_MAX(1, 255) yields 1 ;
  * CYCLE_MAX(1, 4) yields 4 ;
  * CYCLE_MAX(0, 128) is ambiguous and what it yields is undefined.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>