Fix issues with ghosts when inserting.
Ghosts can kill us, so take care of keeping their number small. Also
implement ghosts elimination if last of the series (aka next ptr is NULL).

Map in the reverse order for this very reason (if the map removes an entry,
better remove it at the end of a series to allow not creating a ghost).

*** Needs unit testing. ***

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>