Hashtables where keys are strings.
Make hashtbls work without hacks, dealing with collisions, for types we sort
using one of their member as a name.

To define such a hashtable type, supposing it holds foo_t's and that foo_t
has a member called foo_name that is of type char *, you just need to:

  DO_HASHTBL_STR(foo_t, foo, foo_name, false);

if foo_name was a char[], the last parameter would actually be `true`.

Notes:

There is a cost using those hashes (the strcmp), so when you store things
that have a perfect hash, use the simpler hashtbls that assume no
collisions.

The hashing using code can use whichever hashing functions it wants.  hsieh
is the fastest, jenkins has a slightly better discrimination rate, but if
quality of the hash is seeked, the combined hash (using both previous ones)
has to be used.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>