ghost related refactors.
Define GHOST to (void *)1 and document that their key is completely
uninteresting to us. We're interested in setting a value in its proper slot
directly when possible, but if we are in an overflow case, we already lost,
there is absolutely no gain in reusing a slot used by the same key or not.
For the matter, we could zero-fill the key with the same behavior in the
end.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>