Convert a tpl into an iovec.
No need to have an agent_write_tpl, just build an iovec with it, and consume
it.

The function returns the number of iovec cells that are needed, and doesn't
allocate anything.

I believe callers should have a fixed iovec size (e.g. with 128 cells), and
if it doesn't fit, consume the whole iovec, and when consumed, ask to "drop"
the 128 first cells producers from the template and start over: consuming
tpl nodes enforces too many tpl nodes reallocations when refcounts don't
fit, and that just sucks too badly, we have to work en-masse instead, hence
the proposal.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>