No it's not an error, it's modeled after the python API.
Documentation will come in due time. API is that path_join joins atoms of
paths with the "proper" separator. It has no semantics attached, and is just
a facility.

We may have a more clever function at some point that does what you suggest,
but here we just want that.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>