Don't try to reuse ->skip'ed data if it's for a few octets.
To avoid a scheme were we do push-back, pop-front in cycles, with a size
of n+1, and n elements in it. We would do memmove of lenght n for each
push-back which sucks badly.

Of course rings are a better structure for such uses, but real life
legitimate usages of blobs/arrays/vectors could trigger such issues too.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>