Use memoization in core-debug.c
We memoize the level associated with a given [file,function] for real.

This will allow to have way more complicated stuff in IS_DEBUG, and
have arbitrary complex lookup and constructions like:

lib-common:0 lib-inet:0 lib-inet/agent.c:2 mcg:100

which would disable debug in the lib-common and the lib-inet, but have it
at level 2 in agent.c and level 100 for everything in mcg/. This is just
an example, but the point is we can have an all-matches-are-needed policy
now and do not suffer from it.

We even can use dumb lookup that parse IS_DEBUG each time and use complex
stuff like fnmatch so that we can have globbing in IS_DEBUG which would
quite rock.

Signed-off-by: Pierre Habouzit <madcoder@debian.org>