el: reorganize fields and annotate which type uses them.
Prepare stubs for a new kind of el_t: EV_PROXY.

Thanks to already existing holes in the structure, it doesn't grow the
struct for 64bits pointers, and adds 4 octets to its size on 32bits
platforms.

Add the watched events mask to the el_show_blockers output.
Add the proxy events to el_show_blockers.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>