el: cache lists instead of dlist_for_each_entry_safe.
When you remove the "next" element from the list you're basically screwed
(hard) so we now cache all the results then process. It's slightly slower,
though it's safer, and allow us to unregister directly for every kind of
ev_t and push them into the gc queue.

The gc queue is collected on a regular basis in O(1).

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>