blob_gsm: add some comments.
Algorithms are tricky, and what they do is quite specific to their task.
blob_append_gsm7_packed is not meant to be a really restartable function,
but aims to help implementing UDH+BODY packing for SMS that is a very
specific task.

Explain that, and hint how to do a restartable implementation: just pass
the `septet` internal counter as an inout parameter set to 0 at init time.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>