rework license code for cleaner integration.
Instead of the ugly "let's live in macros.h" approach, let's live in
el-showflags.c and be hooked into the even loop, completely inlined to
avoid "let look the function address" approach.

We even force the inlining with always_inline so that we're really sure
gcc doesn't screw with us on this.

There is a subtle hack to be sure license and check-license come together:
the trace_override is used by the trace code _and_ put in license.c, so
that the modules depends on each other.

As a consequence core-macros.h goes uncluttered.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>