Update our compulation flags.
Inlude -finline-functions to inline local static short functions when
`inline` is not specified (our coding rules forbid inline in local static
functions so we need this).

use -fpredictive-commoning and -funswitch-loop to be able to optimize
loops wrt invariants. The latter especially does that:

                                  if (a & 1) {
for (;;) {                            for (;;) {
    if (a & 1) {                          do_this();
        do_this();                    }
    } else {            ---->     } else {
        do_that();                    for (;;) {
    }                                     do_that();
}                                     }
                                  }

When of course a is not used inside the loop. It allows to chose concision
over speed for the code, and still produce efficient code.

finally -funsafe-loop-optimizations let gcc assume that when you write:
    for (; i < limit; i += 4)
the loop won't overflow. (IOW that if i and limit are uint8_t, limit isn't
e.g. 255). It allows gcc to perform some further optimizations it seems,
and I see no reason to avoid that.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>