lib-common: rework intersec memory wrappers.
This will in the end allow memory allocated through pools to be handled in
a generic way. Of course, one doesn't really want to realloc/free memory
with the generic wrappers when one knows where the memory comes from. But
it's now possible to do so (in theory at least) with not too silly
performance.

This commit migrates the mem-fifo-pool to this new scheme. The overhead
for the generic case is minimal (registering and unregistering into the
memory rb tree). Though because of the xml abusing the model to use it as
an obstack, the current mem_fifo_pool_delete code is clearly suboptimal
(we walk through all the registered memory blocks to find our own and
collect them). Luckily, in the usual case, the xml mem fifo pool has only
one page: the current one, so the code that actually walks through all the
memory blocks should not be triggered often.

TODO: write an actualy obstack-like pool

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>