memory: various updates.
Drop mem_* macros and use straight i*alloc calls.

Enhance irealloc inlining for the case where the compiler knows if
oldsize < size.

Enhance free inlining when we statically know the memory is NULL.

We don't care if oldsize is 0 in irealloc, really. irealloc means malloc
as soon as _*mem is NULL_.

specify that (size_t)-1 means 'unknown size' for oldsize in irealloc.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>