stacked pool: document it a bit, fixes, enhancements.
Make the `last` optimization cached into the frame, it helps when you are
reallocating your last value, then call a function that uses the stack
temporarily via push/pops, you really want your further reallocs to be
efficient.

fix align_boundary that has to compute the boundary at the biggest 2 power
that is smaller or equal to the size, and not the smallest strictly bigger
one. (IOW make align_boundary(3) return 2 instead of 4).

fix blk_destroy that stupidly missed the ifree() call.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>