el: force get_clock when we have no timers.
get_clock is refreshed regularly only if we have timers, so we have only
one timer in the heap (or none) force a refresh each time we register one.

The 1-oddity is to support cases where we're not sure if the timer we're
registering was there at the last refresh. Starting at 2, we're sure the
other one triggered a refresh recently enough.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>