htbl: use double-hash linking instead of linear probing.
On collision, we previously looked in a linear fashion for empty cells.
Now use double hashing, meaning that depending on the starting cell, we
don't search for the next cell in the same fashion. This reduces lengths
of collision chains greatly.

Now that our htbl sizes [size] are primes, for a hash of [h], the first
cell we look at is [h % size] and the increment used is
[1 + h % (size - 1)]. size being prime, we know that this increment allow
a full scan of the h-table.

As a consequence, we cannot purge trailing ghosts in htbl_invalidate
anymore.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>