Fix stack push
On push we allocated the new frame structure _inside_ the previous frame
allocation space. And on pop, we still had this structure on top of the
frame.

Hence for a given push()/pop() cycle, a frame_t was "leaked"...

We now split the allocation core function of the mem stack into:
- sp_reserve which only reserves the space for the allocation, prepare
  blocks, and returns the proper slot pointer ;
- sp_alloc which effectively update the owning frame with this allocation.

push() now only uses reserve instead of sp_alloc...

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>