Fix snprintf to never write in the place where it puts the \0
Else if we do that:

    sb_t sb1, sb2;

    sb_init(&sb1);
    sb_init(&sb2);

    sb_addf(&sb1, "+%s", sb2.data);

THen sb1 ends up containing "+\0" instead of "+" because for a short
while, '+' that is the first char of the format gets written to the
__sb_slop that both sb1 and sb2 share just after the init.

Hence when the strlen of sb2.data is evaluated, it's strlen "+" that is
evaluated. snprintf says it requires 2 octets, then the second run after
that the sb1 buffer is grown doesn't suffer from the issue and only "+" is
written as expected. sadly we trust the first printf and have allocated 2
octets.

Of course, this isn't strictly speaking a bug in sprintf because we pass
an output buffer that is also an input one. but the previous pattern is
obviously something that ought not to fail this way because this is
totally hidden to the programmer.

FWIW I believe it fixes the bug that makes us need to make the __sb_slop a
TLS variable...

Signed-off-by: Francois Lodier <francois.lodier@intersec.com>