Add filtering on a user-defined name for tracing.
This is useful to have the traces for a feature and not only for a file or
a function.

You can use e_named_trace to assign a feature name to a trace log. Then
you can get all the traces for a given feature name by using the '+'
modifier in IS_DEBUG.

e.g.:
    e_named_trace(2, "my_feature", "My trace: %s", blah)
    IS_DEBUG="*+my_feature:2"

Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>