qh: fix possible overflow in qhash_scan.
{min,max}size may not be a real hash size when we're dealing with old,
since we use qh->old.len (which is the last pos at which we have not yet
moved datum) instead of qh->old.size as an optimization to stop the scan
earlier.

Sadly, the conditions to exit the scan loops used the fact that we have a
guard bit at "{max,min}size" exactly, which may never happen if the
{max,min}size is actually a qh->old.len (since there isn't a guard bit at
that position).

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>