Fix tlsf allocators on "realloc"
realloc implementations forget about the reallocated block "PREV_FREE" bit
when realloc was able to reuse the same block.

This breaks tlsf invariants badly: if the bit is set, then when *this*
block will get free()d it won't be merged with its previous block, and
tlsf heavily rely on the fact that there isn't two contiguous free blocks.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>