clstr rework.
This is a deep change, clstr_t are no more, lstr_t track whether the
string is allocated (and where).

This intends to get rid of iop_string_{dup,free,...} horrors in favor of:
- lstr_dup/lstr_copy to allocate strings from the libc
- l_lstr_* to allocate strings from the mem stack
- mp_lstr_* to allocate strings from a given mem pool
- lstr_wipe to deallocate the string (or do nothing when it's constant).

lstr_t are supposed to be immutable, meaning that writing into lstr_t::v
should be restricted to specific needs and avoided as a general rule.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>