thr-job: thr_syn_wait now spins on the event count.
It allow us to make the thread progress (steal jobs and similar) when
we're stuck on a slow task since nobody can unblock us (but for the the
people knowing about the syn ec).

Also in the main thread, if there is work on the main queue go back see if
there is something on our queue again.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>