t_stack realloc checks: fix 6344858
the stack frame bookkeeping must be done in sp_alloc and not sp_reserve.
The former is the straight implementation of t_new whereas sp_reserve is
internal (and is used to allocate frames e.g.).

In addition to that, realloc should just check for the frame but not
include the pointer else it breaks when the realloc actually moves memory.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>
Reported-by: Jean-Marc Coïc <jean-marc.coic@intersec.com>