thr_syn_wait: rework queueing in the main queue
The main thread can only wait on thr_syn_t's, make it advertise which one
so that people queueing in the thr_main_queue:
(1) do not call pthread_kill for nothing
(2) unlock the proper event count.

This makes waiting for the main thread more expensive when it hits the
slow path, which we don't care.

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>