el: add the idle event
Having active idle events makes the loop non blocking (using a 0 timeout
in epoll).

If the loop then does nothing (but for before hooks) then the idle events
are called.  Once an idle event has been called it is parked.

The system decides to unpark parked idle watchers when:
- the epoll has blocked for a long time (100ms).
- the last idle watch event has been fired more than 10 minutes ago.

The second case is to be able to still run "idle" tasks on a relatively
busy system (e.g. something that interacts every 50ms or so, but with a
low CPU usage).

Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>
Acked-by: Florent Bruneau <florent.bruneau@intersec.com>