iop-xml: fix WSDL generation of the IOP xml type
When the XML type was used in an IOP object we converted it to the XML
type xsd:any to say it's a piece of valid XML and that's all. But this
isn't sufficient for WDSL validators which expect the content to be valid
regarding to the current namespace.

So we have to add the attribute processContents="skip" to say to the
validators to not try to check the XML with the current namespace.

See: http://www.w3.org/TR/xmlschema-1/#Wildcard_details

Closes: #4969
Reviewed-by: Pierre Habouzit <pierre.habouzit@intersec.com>