thr-job: fix deadlock when interleaving several thr_syn_wait
If the main thread steal a job which does a thr_syn_wait, it will destroyed
the thr0_cur_ec_g variable. And then the atomic_bool_cas will never work
and we will deadlock.

To fix this, we forbid the main thread to steal jobs. This is not the more
efficient way (we lost a thread to work) but it seems to us the safest
solution.

Refs: #5487
Refs: #5514

Reviewed-By: Florent Bruneau <florent.bruneau@intersec.com>
Acked-by: Pierre Habouzit <pierre.habouzit@intersec.com>