http(d): fix queries reference counting.
This fixes two bugs with http machines and IOPs

(1) IOPS: obj_delete are just wrong, it shouldn't be done.
    I don't think it really affects anybody right now (except for igloo
    zpush test daemon :P) as you need to have an IOP HTTP Daemon serving
    IOP async RPCs, which we don't.
(2) reference counting of httpd_queries is plain wrong, as a consequence
    when the client of an httpd quits, crashes can occur because we delete
    the query instead of keeping it live.

    This is only a problem if you're answering asynchronously though.

The fix consist in using the objects reference counters, and have 1 unit
for the fact that we're attached to a connection, and 1 unit for the fact
we've not been answered to yet.

Closes: #5796
Cc-Redmine: yes
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>
Acked-by: Jonathan Squirawski <jonathan.squirawski@intersec.com>
Tested-by: Jonathan Squirawski <jonathan.squirawski@intersec.com>