obj: make refcnt a ssize_t and all negative values magical.
in retain:
 - 0 should not happen, as it should be reached only to mean "destroy this
   object". It's likely we're retain()ing something that was just
   releas()ed and that valgrind is about to complain loudly.

 - if the refcnt is 0x7fff....f then we're overflowing (unlikely but...)

 - all other values are forbidden for now

in release:
 - if we're at 0, we're release()ing too much
 - at 1 we dealloc()
 - all other values are forbidden.

The change wrt before that may have an impact is that it used to be
allowed to wipe a "static" object with obj_release() or obj_wipe(), which
nobody does.  Now we need it to be obj_wipe().

as an extra security, ensure in obj_wipe() that the refcnt is either
negative or 1.

Change-Id: I9c418b39b6e8449a660f3441934f6d059764bf9d
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>