http: 404 for unresolvable triggers if there is a trigger for the method.
When the http layer was written, it was meant that people would always
register a "/" trigger acting as a catchall that would 404 on HTTP methods
we implement, hence would "501" unused methods (PUT, DELETE, ...).

In the reality, we never do that catchall, and get 501 on GET/POST which
is, err, curious at the very least.

This patch turns that into a 404 when there is at least one trigger
registered on the given method meaning that the *method* itself is
*implemented* but that nothing is there to answer.

We will still 501 GETs if the http server only answers to POSTs, that's
why we now have a "no handler for [POST,GET,...]" which helps debugging.

Change-Id: Id8d4660541c42fbc89cd94de0a0de97f78138b86
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>