xmlr: make the "nullok" semantics more logical.
Rename the field as emptyok, meaning that the TEXT value of the node is
allowed to be empty or not.

Also when the node is empty, return LSTR_EMPTY_V rather than LSTR_NULL_V
which makes more sense.

I've checked all the callsites with {nullok = true}, callers protects
themselves against the NULL-ity but check 'len' to know if the value is
empty, hence this change should be safe.

Change-Id: Ie8a1aa7456fef15c30c263eab46bcb1d0c66a2a3
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>