iop-xml: fix handling of IOP_XUNPACK_IGNORE_UNKNOWN flag
When xunpack_struct is called on an empty node, xmlr_next_child fails with
XMLR_NOCHILD. In that case, we do not want next_uncle but node_close,
since we did not enter in the node.

Besides when we are in the end label, we have either is_closing or
is_empty so node_close is the right thing to do (probably the first bogus
version of my patch broke that assumption).
Anyway when there are unknown tags at the end of a sequence (after all the
known tags), we are actually in the case 'fdesc == end', and that's where
we want to call next_uncle.

Refs: #5991
Change-Id: I9158834e0ffde6ca8a5001358cd035c89d5bb821