Z: add support for "modes".
We currently have only a single mode: "fast".

Tests can alter their behavior according to the activation (or not) of
modes. For example, when mode "fast" is activated, you can reduce the
number of element of your datasets in order to speed up the processing.

Modes can be set using the Z_MODE environment variable or through the
command line argument with --z-mode (for the C runner only). Z_MODE is a
list of space separated modes.

Change-Id: I9aeb69a62e1cfd20de113aa4f371b48a0bcbc77a
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>