lib-common: get rid of the silly mem_blk_t feature.
This was intended for containers and such to be allocated on pools
transparently but isn't used, and adds considerable overhead in our
allocators.

Just drop that feature.

Note for readers: mem_blk_t structure lived in core-mem.h and was removed
and its member 'start' and 'size' have been inlined in the embedders, its
rb_node_t and pool members removed as those were used for the feature
described before.

Change-Id: I39fcb0a4aa441fec2d71d77c0f20e40509f7ce18
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>