http: do not obj_release if parsed && !answered.
Being parsed means we've called on_done, hence the responsibility to
answer the query has been handed off to the application. So when it does
(which it has to), obj_release()ing when yet unanswered is one release too
many.

When we're not parsed, we haven't called on_done yet, and the query is the
http layer responsibility, hence it has to obj_release. We may be !parsed
&& answered (think cases when we're waiting for the end of a request that
has been rejected after having read Headers for example).

Closes: #6382
Change-Id: I3242b331ef8d0eb3b029126ad89797b5ad21923d
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>