iop: optimize the IOP packing/unpacking of the “small” repeated fields
Currently the “small” repeated fields (byte[], ubyte[], bool[], short[],
ushort[]) are packed in an inefficient manner. All these types should be
packed like the type “bytes”.

This commit transform these types into “bytes” in a transparent way. These
types are now encoded as “bytes” and the unpacker keeps the ability to
unpack the old format.

Be careful that this will break compatibility of the IOP data in
case of rollback.

Closes: #6561
Change-Id: I4ebea37ac3d3e92630a4c72038ffc8cc368be380