iop: avoid to pack default values for nothing
When a field is equal to its default value it's useless to pack. It uses
bandwidth for nothing and the unpacker will do the right thing if the
field is absent (i.e. set the default value).

Be careful that we do a very simple comparison of values equality to be
efficient. This especially true for “string”, “bytes” and “xml” types. We
consider a string equal to its default value only if its data pointer is
the same as the default value one (set by iop_init) and if their lengths
are equal. The string content isn't compared with the default value one.

We handle another special case. If a “string”/“bytes”/“xml” is equal to
LSTR_NULL_V then we considerer this as the default value. So a p_clear on
a string member will do the thing that most users expect.

Closes: #4351
Change-Id: I04e3164ab9f6df161e8240997c211c9176bda632
Cc-Redmine: yes