iop: allow to easily forward the answer from one RPC slot to another
We often want to do IOP proxies but manual one to do some checks before to
profixy the query. This force the user to manually transfer the RPC answer
to the initial client even if we don't care about the response.

The new “ic_forward_reply_to” allows to transfer the answer of an RPC to
another RPC slot. This API saves some code and transfer the answer more
efficiently because the IOP object isn't re-packed. We just forward the
received binary stream.

Yet, this API is somehow dangerous. So if you use it, do it carefully. The
API considered that you HAVEN'T MODIFY the RPC answer. If you try to do
this, in the best scenario your chances will be ignored, in the
worst you will have a crash…

Here an example:

RPC_IMPL(pkg, foo, bar)
{
    CHECK_ID_OK(arg->id);

    /* manual proxy */
    ic_msg_t *imsg = ic_msg_new(sizeof(uint64_t));
    *(uint64_t *)imsg->priv = slot;
    ic_query_p(remote_ic, imsg, pkg, foo, bar, arg);
}

RPC_CB(pkg, foo, bar)
{
    uint64_t origin_slot = *(uint64_t *)msg->priv;

    /* automatic and efficient answer forwarding */
    __ic_forward_reply_to(origin_slot, status, res, exn);
}

Closes: #4412
Cc-Redmine: yes
Change-Id: I7aced16d2a0654aa0f63003806e195e023f8adea