qheap: fix position setter for inlined types
Problem:
~~~~~~~~
Into __qhp_##n##_set, we used to set qhp vector element before updating
position. It is OK when the element is a pointer because we always update
the right object, but if the element is inlined, we just update a copy of
the element, not the element itself.

Solution:
~~~~~~~~~
Update position on the local variable *then* copy it into the qheap vector.

Note:
~~~~~
This patch also add unit tests to raise the error. We do not really expect
anybody to store position if the contained objects are inlined but it has to
work.

Refs: #6670
Change-Id: I48ee838d7b31bf48c9bf6272ebaf4eee8e545915
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>