thr_syn_t: make it safe wrt multiple waiters.
Right now, waiting on a thr_syn_t if you're not the owner is just plain
broken, and subject to races when the thr_syn_t is for example on the
stack of its owner.

We used the "unsafe" technique to mitigate that, but it was a poor hack,
here is a more complete refcounting based solution.

Now when someone waits or owns the thr_syn_t it retains the thr_syn_t
until the wait is over or it decides to destroy the thr_syn_t.

A sched_yield() is added in the eventcount because now when you're awoken
after a futex call there can be a herd effect, the sched_yield mitigates
it trying to avoid the contention on the thr_syn_t::ec::waiters then
thr_syn_t::refcnt atomic integers.

Given that there are very few waiters (at least one, sometimes two, rarely
more) this overhead is totally acceptable and should in fact yield almost
no contention in the real world.

NOTA: now it's *NECESSARY* to do a thr_syn_wipe() else your code will
crash.

Change-Id: If25593c29146b5a0ff985e8829146cd74ac99ca7
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>