eventcount: shift the burden of updating "waiters" to the signaller.
The guy calling thr_ec_signal_n is doing atomics in the thr_evc_t. Letting
the awaking thr_evc_t decrement their waiting count generates contention
in the case of broadcast and that's silly.

In the case when we're woken up, assume the waker will decrement the
waiter count for us. Do it in all the other cases.

Also return the number of threads that have been woken up.

Change-Id: Id3afcffb2ee310d96620492bfbf2c84a76ea339a
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>