thr-evc: make futexes private
See linux-2.6.git 34f01cc1f51 for an explanation.

A PRIVATE futex is a futex that is private to the process, and not
system-wide. This is the same distinction between a
PTHREAD_PROCESS_{PRIVATE,SHARED} mutex (see
pthread_mutexattr_setpshared(3p)).

We only use eventcounts locally to a process, so gain the extra bit of
performance we can have here.

Change-Id: I011de2b43c3ae90544220044574603b585fd2d09
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>