mem-fifo-pool: several fixes.
This patch several confusions between the size of the block and the size
of the allocated area. This led to a buggy valgrind instrumentation and
more seriously could cause a buffer overrun in the reallocator (when
resetting the value to 0).

Also fix the ordering of the operation in mfp_free that could lead to the
leakage of the current page when deletion of the pool was postponed (aka
!mfp->alive).

Refs: #6984
Change-Id: Ib99b064e5de6002f2d0b7e2e78bf999f86d7c088
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>