qps: on recycling older TLSF maps, shrink them
Else we may occupy more disk than expected. Also madvise(DONT_NEED) the
pages to help the kernel.

The file is supposed to be here, hence we don't deal with errors. Just
truncate it to a page then back to a map size so that it's sparse and full
of zeroes again.

Change-Id: Id9c248fdba9ec54fe3327101a0503c11615d8bbe
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>