httpd: add notification callbacks for connections/disconnections/query termination
on_accept is called after the httpd has been spawned in order to ensure we
have access to informations such as the configuration or the file
descriptor (useful if you want to log the remote address).

on_status is called each time a query's lifecycle ends. We ensure the callback
has access to description of the query (such as the method, the path, ...)
by setting the q->qinfo while parsing.

Refs: #6751
Change-Id: If53624be8c05dea9ed6ff630ee7449590b80e906
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>