el: add support of priority on fd.
el_fds can have 3 differents priorities: LOW, NORMAL or HIGH, default is
NORMAL. After a poll, on the events of the fds with the higher priority
are performed (that is, NORMAL/LOW events are never performed if at least
one HIGH fd has event to be processed).

This patch includes a minimal testcase.

Closes: #6733
Change-Id: I1080a322ff919e564ad3464463bbd3a79910ac02
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>