http: fix regression introduced by cdf624c2d65b.
In order to be able to access the details of the query within the
notification callbacks, q->qinfo is set to a pointer to httpd_qinfo_t
structure allocated on the stack. Since the pointer was left untouched at
the end of the function, this led to freeing a pointer on the stack when
the httpd_query_t is release.

Change-Id: Ib32e0e01f11b44f46c4ddfff263a116d77fccfa8
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>