qps: introduce x_finalize_meta
We call qps_map_m_recycle *before* the meta.qps is actually renamed, which
is WRONG. qps_map_m_recycle is cleanup, and should be done afterwards.

This is the real source of the m.disk_usage == 0 issue.

x_finalize_meta does the rename then the cleanup, in the proper order.

Change-Id: I4cccfa53900b1e1affcb9b33936afcb44d80937f
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>