http : Modify httpc_query_start to not urlencode
Modify httpc_query_start function in order to let the developper to
encode only a part of the url. In this case,the function do not encode
the url anymore.

For backward compatibility, we add a define that match the previous
definition of the fucntion and rename it on httpc_query_start_flags.
Then a fifth arguments has been added in orderto control the behavior
of the httpc_query_start_flags function.

Change-Id: I50611b17585f3ae49519872fba2c30e8619a7ab3
Signed-off-by: Julien Richefeu <julien.richefeu@intersec.com>