qhat: fix 'safe' mode of qhat_tree_enumeration_go_to.
We properly handled the case of structural changes in the trie but not
minor changes within a compact node. So, ensure cached values of en->count
and en->key are valid before jumping to the required value.

This is a regression introduced by the refactoring of the qhat enumerator
and thus only affects branches that uses qhat from the lib-common.

This is the real fix for the underlying bug of change Ie1b7c026.

Change-Id: If1c887fc67abbd834339f2e2ad557e53545215a3
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>