http[cd]: support gzip/deflate transfer-encoding on "receiver" end.
Meaning in queries for the httpd case, and on answers for the httpc case.

Since we support that, always generate an accept-encoding header allowing
gzip/deflate as a client (we do not support compress since zlib doesn't).

This merely requires a struct z_stream in the httpd/httpc objects, and
hook in the data flush code. Nice and easy.

This patch has been tested manually with gzipped files and very
small-grained chunked encoding of those using curl/POST by hand (urgh).

Efficiency note:
~~~~~~~~~~~~~~~
  We're not very copy-efficient and copy data around 3 time:
  - gzipped data is read() into the ibuf,
  - then data is decompressed in the temporary zbuf,
  - it's likely that data then is copied to the "payload" sb in
    http[cd]_query_t.

  IOW we copy uncompressed data twice, and compressed data once. Too bad.

  In *THEORY* using inflateBack could help, though we have to know way too
  much about gzip/zlib/... for my taste just to save a few memcpy's

V2:
  use inflateReset2 to reset to avoid a malloc/free cycle.
  fix typo in error message.

Refs: #6734
Change-Id: I5e8be64a54ebad7bf0c27df4640c76cedf9681c7
Cc-Redmine: yes
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>