iop: http: add support for gzip.
When the client indicates it accepts gzip, compress the reply.

The do { } while () my look expensive, but the probability the
compressed data is larger than the original one is very small (it should
only happen when the original content is very small, in which case we
could simply decide not to compress the data).

The compression has a double cost: it implies an extra copy of data and
increases CPU usage but on both JSON and SOAP this should have a huge
impact on the size of the payload (early tests show a decrease of 40 to
70% of the size of the data in medium-sized SOAP requests).

Refs: #6734
Change-Id: I8fb64a5e481b54272b431052c4d4112b17785621
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>