qps: map the pages in ANONYMOUS pages.
To do that we now have an "omaps" (as old maps) array of old map addresses
we can reuse. When recycling a page, the adress is kept around in that
array.

When mapping a file, we look for an already existing available map adress,
if it exists, we reuse that.

Unused files are now unlinked from the disk directly.

V2: Reinstate deleting the file before we map it, it's safer, we'll get a
    new inode this way.

V3: Richard is right, MAP_SHARED | MAP_PRIVATE makes no sense.
    This means though that we could now snapshot with a fork() which would
    greatly simplify the snapshotting, allowing us not to require the
    SIGSEGV handlers… Or at least in a more efficient way…

Change-Id: Ia315829617405c370da8d4004e03d7a09cd01c71
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>