qps: implement a fork-based-COW snapshotter
Now that the pages are in MAP_PRIVATE maps, we can use the MAP_PRIVATE
semantics to implement snapshotting.

No more malloc() in the sigsegv handler, code is significantly smaller, we
can gzip better since we won't lock maps for long anymore, and finally we
can nice() the snapshot to have less impact as well.

Change-Id: Ie84390b8fc138e8f2e0c1c8e5271429774aaa093
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>